Improve fonts.conf to assign directly to our fonts
Use the same strategy as the GTK and Chromium port and assign directly the
fonts used in LayoutTests to our available font choices. Also remove rules
that do not help this goal.

The fonts-qt5-wk1.conf is not needed anymore because the direct rules work
for Qt 4.8.2 and Qt 5.

Use Liberation font family for tests, instead of Nimbus.
Liberation provides the same metrics as MS Core fonts, and is used by
other ports, for example GTK.

The patch also fixes the problem of returning the wrong fonts when
monospaced was request.

The old expected png results were not changed, and will be properly
updated during the unskip stage.

For further information on the font discussion, see the webkit-qt mailing
list thread[1] and bug #85203.

[1] https://lists.webkit.org/pipermail/webkit-qt/2012-May/002846.html